Add files via upload
Jan 8th 18:30 - 20:20
Just a side note, I just realized that the question I worked on for the previous session was #5, not #6. For this session, I completed three problems. #6 went fairly smoothly except I spent 10 minutes trying to figure out why my function kept giving me a result of 1. It turned out that I indented too much for the return statement. Stupid mistake, but important to remember the lesson. Understanding the biological terms in #7 took some time. The logic and math of #7 was kind of complicated, so I drew some graphs to help myself understand. Once the math worked out, it became easy to complete. For #8, I went back to lab 2 for some inspiration. Typing the codon translation took very long.